requirements.txt
streamlit
scikit-learn
spacy
PyPDF2
python-docx
altair